WIP refactor into a class
