+++ constants.py
@@ -1,0 +1,117 @@
+from enum import IntEnum, IntFlag
+
+STATUS_OFFSET_ERROR_INFORMATION_1 = 8
+STATUS_OFFSET_ERROR_INFORMATION_2 = 9
+STATUS_OFFSET_MEDIA_WIDTH = 10
+STATUS_OFFSET_MEDIA_TYPE = 11
+STATUS_OFFSET_MODE = 15
+STATUS_OFFSET_MEDIA_LENGTH = 17
+STATUS_OFFSET_STATUS_TYPE = 18
+STATUS_OFFSET_PHASE_TYPE = 19
+STATUS_OFFSET_PHASE_NUMBER = 20
+STATUS_OFFSET_NOTIFICATION_NUMBER = 22
+STATUS_OFFSET_TAPE_COLOR_INFORMATION = 24
+STATUS_OFFSET_TEXT_COLOR_INFORMATION = 25
+STATUS_OFFSET_HARDWARE_SETTINGS = 26
+
+
+class ErrorInformation1(IntFlag):
+    NO_MEDIA = 0x01
+    CUTTER_JAM = 0x04
+    WEAK_BATTERIES = 0x08
+    HIGH_VOLTAGE_ADAPTER = 0x40
+
+
+class ErrorInformation2(IntFlag):
+    WRONG_MEDIA = 0x01
+    COVER_OPEN = 0x10
+    OVERHEATING = 0x20
+
+
+class MediaType(IntEnum):
+    NO_MEDIA = 0x00
+    LAMINATED_TAPE = 0x01
+    NON_LAMINATED_TAPE = 0x03
+    HEAT_SHRINK_TUBE = 0x11
+    INCOMPATIBLE_TAPE = 0xFF
+
+
+class Mode(IntFlag):
+    AUTO_CUT = 0x40
+    MIRROR_PRITING = 0x80
+
+
+class StatusType(IntEnum):
+    REPLY_TO_STATUS_REQUEST = 0x00
+    PRINTING_COMPLETED = 0x01
+    ERROR_OCCURRED = 0x02
+    TURNED_OFF = 0x04
+    NOTIFICATION = 0x05
+    PHASE_CHANGE = 0x06
+
+
+class PhaseType(IntEnum):
+    EDITING_STATE = 0x00
+    PRINTING_STATE = 0x01
+
+
+class PhaseNumberEditingState(IntEnum):
+    EDITING_STATE = 0x0000
+    FEED = 0x0001
+
+
+class PhaseNumberPrintingState(IntEnum):
+    PRINTING = 0x0000
+    COVER_OPEN_WHILE_RECEIVING = 0x0014
+
+
+class NotificationNumber(IntEnum):
+    NOT_AVAILABLE = 0x00
+    COVER_OPEN = 0x01
+    COVER_CLOSED = 0x02
+
+
+class TapeColor(IntEnum):
+    WHITE = 0x01
+    OTHER = 0x02
+    CLEAR = 0x03
+    RED = 0x04
+    BLUE = 0x05
+    YELLOW = 0x06
+    GREEN = 0x07
+    BLACK = 0x08
+    CLEAR_WHITE_TEXT = 0x09
+    MATTE_WHITE = 0x20
+    MATTE_CLEAR = 0x21
+    MATTE_SILVER = 0x22
+    SATIN_GOLD = 0x23
+    SATIN_SILVER = 0x24
+    BLUE_D = 0x30
+    RED_D = 0x31
+    FLUORESCENT_ORANGE = 0x40
+    FLUORESCENT_YELLOW = 0x41
+    BERRY_PINK_S = 0x50
+    LIGHT_GRAY_S = 0x51
+    LIME_GREEN_S = 0x52
+    YELLOW_F = 0x60
+    PINK_F = 0x61
+    BLUE_F= 0x62
+    WHITE_HEAT_SHRINK_TUBE = 0x70
+    WHITE_FLEX_ID = 0x90
+    YELLOW_FLEX_ID = 0x91
+    CLEANING = 0xF0
+    STENCIL = 0xF1
+    INCOMPATIBLE = 0xFF
+
+
+class TextColor(IntEnum):
+    WHITE = 0x01
+    OTHER = 0x02
+    RED = 0x04
+    BLUE = 0x05
+    BLACK = 0x08
+    GOLD = 0x0A
+    BLUE_F = 0x62
+    CLEANING = 0xF0
+    STENCIL = 0xF1
+    INCOMPATIBLE = 0XFF
